raise tmp

+++ include/linux/kvm_dirty_ring.h
@@ -55,7 +55,7 @@ static inline int kvm_dirty_ring_reset(struct kvm *kvm,
 }
 
 static inline void kvm_dirty_ring_push(struct kvm_vcpu *vcpu,
-				       u32 slot, u64 offset)
+				       u64 slot, u64 offset)
 {
 }
 
@@ -87,7 +87,7 @@ int kvm_dirty_ring_reset(struct kvm *kvm, struct kvm_dirty_ring *ring);
  * returns =0: successfully pushed
  *         <0: unable to push, need to wait
  */
-void kvm_dirty_ring_push(struct kvm_vcpu *vcpu, u32 slot, u64 offset);
+void kvm_dirty_ring_push(struct kvm_vcpu *vcpu, u64 slot, u64 offset);
 
 bool kvm_dirty_ring_check_request(struct kvm_vcpu *vcpu);
